split Server.hpp
